fix: split exception classes into separate files - Create separate files for each exception class - Fix autoloading issues - Maintain proper class inheritance
